# rest/api/swagger.yaml
---
swagger: "2.0"
info:
  title: Oshinko REST Server
  description: |-
    The REST API server for the Oshinko suite of applications
  version: "0.0.0"
consumes:
  - application/json
produces:
  - application/json
schemes:
  - https
  - http
definitions:
  ClusterModel:
    allOf:
      - $ref: "#/definitions/NewCluster"
      - properties:
          status:
            description: |-
              Current status for the cluster
            type: string
          masterUrl:
            description: |-
              URL to the spark master
            type: string
          masterWebUrl:
            description: |-
              URL to the spark master web UI
            type: string
          pods:
            description: |-
              Pods that make up the cluster
            type: array
            items:
              type: object
              properties:
                ip:
                  description: |-
                    IP address of the pod
                  type: string
                type:
                  description: |-
                    Type of pod
                  type: string
                status:
                  description: |-
                    Pod status
                  type: string
              required:
                - ip
                - type
                - status
        required:
          - status
          - masterUrl
          - masterWebUrl
          - pods
  ErrorModel:
    properties:
      status:
        description: |-
          HTTP status code
        type: integer
        format: int32
      title:
        description: |-
          A brief summary of the error, this should remain consistent
          between errors of the same class
        type: string
      details:
        description: |-
          An explanation of the specific error that has occurred
        type: string
    required:
      - status
      - title
      - details
  ErrorResponse:
    properties:
      errors:
        description: |-
          List of errors returned
        type: array
        minItems: 1
        items:
          $ref: "#/definitions/ErrorModel"
    required:
      - errors
  NewCluster:
    properties:
      # INFO(elmiko) my thoughts on the name property is that we can use
      # the naming guidelines defined by kubernetes[1] for labels to help
      # inform our process here. i am thinking that we could use the name
      # as a label on all cluster objects, this will make finding all the
      # pods/controllers/etc easier.
      # [1]: http://kubernetes.io/docs/user-guide/labels/
      name:
        description: |-
          Unique name for the cluster
        type: string
      config:
        description: |-
          Cluster configuration values
        type: object
        properties:
          name:
            description: |-
              The name of a stored cluster configuration
            type: string
          masterCount:
            description: |-
              The count of master nodes requested in the cluster (must be > 0)
            type: integer
            format: int
          workerCount:
            description: |-
              The count of worker nodes requested in the cluster (must be > 0)
            type: integer
            format: int
          sparkMasterConfig:
            description: |-
              The name of a ConfigMap which holds spark configuration files for the master
            type: string
          sparkWorkerConfig:
            description: |-
              The name of a ConfigMap which holds spark configuration files for the workers
            type: string
          sparkImage:
            description: |-
              The docker pull specification of a custom image to use for spark pods
            type: string
    required:
      - name
    additionalProperties: false
  SingleCluster:
    properties:
      cluster:
        allOf:
          - $ref: "#/definitions/ClusterModel"
paths:
  /:
    get:
      description: |-
        Returns information about the server version
      operationId: getServerInfo
      tags:
        - server
      responses:
        "200":
          description: |-
            Server info response
          schema:
            properties:
              application:
                properties:
                  name:
                    description: |-
                      Application name
                    type: string
                  version:
                    description: |-
                      Application version
                    type: string
                  web-url:
                    description: |-
                      Oshinko Web URL
                    type: string
                  web-service-name:
                    description: |-
                      Oshinko Web Service Name
                    type: string
                required:
                  - name
                  - version
                  - web-url
                  - web-service-name
              # INFO(elmiko) should we also add a versions object?
              # will we want to use a hard version in our URLs, for example
              # /v1/clusters
              # if so, we should add a versions object here to display the
              # relative URLs and information about the versions.
            required:
              - application
        default:
          description: |-
            Unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /clusters:
    get:
      description: |-
        Returns all clusters that the user is able to access
      operationId: findClusters
      tags:
        - clusters
      responses:
        "200":
          description: |-
            Clusters response
          schema:
            properties:
              clusters:
                type: array
                items:
                  properties:
                    name:
                      description: |-
                        Name of the cluster
                      type: string
                    href:
                      description: |-
                        URL for the cluster information page in oshinko-rest
                      type: string
                    status:
                      description: |-
                        Current state of the cluster
                      type: string
                    workerCount:
                      description: |-
                        Number of worker nodes in the cluster
                      type: integer
                    masterUrl:
                      description: |-
                        URL to the spark master
                      type: string
                    masterWebUrl:
                      description: |-
                        URL to the spark master web UI
                      type: string
                  required:
                    - name
                    - href
                    - status
                    - workerCount
                    - masterUrl
                    - masterWebUrl
        default:
          description: |-
            Unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: |-
        Create a new cluster
      operationId: createCluster
      tags:
        - clusters
      parameters:
        - name: cluster
          in: body
          description: Cluster to create
          required: true
          schema:
            $ref: "#/definitions/NewCluster"
      responses:
        "201":
          description: |-
            Cluster create response
          headers:
            Location:
              description: |-
                URL of the cluster detail page within the oshinko rest server
              type: string
          schema:
            $ref: "#/definitions/SingleCluster"
        default:
          description: |-
            Unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /clusters/{name}:
    parameters:
      - name: name
        in: path
        description: |-
          Name of the cluster
        required: true
        type: string
    get:
      description: |-
        Return detailed information about a single cluster
      operationId: findSingleCluster
      tags:
        - clusters
      responses:
        "200":
          description: |-
            Cluster detail response
          schema:
            $ref: "#/definitions/SingleCluster"
        default:
          description: |-
            Unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: |-
        Delete the specified cluster
      operationId: deleteSingleCluster
      tags:
        - clusters
      responses:
        "204":
          description: |-
            Cluster deletion response
        default:
          description: |-
            Unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: |-
        Update the specified cluster
      operationId: updateSingleCluster
      tags:
        - clusters
      parameters:
        - name: cluster
          in: body
          description: |-
            Requested cluster update
          required: true
          schema:
            $ref: "#/definitions/NewCluster"
      responses:
        "202":
          description: |-
            Cluster update response
          schema:
            $ref: "#/definitions/SingleCluster"
        default:
          description: |-
            Unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"